requirements.txt
google-cloud-bigquery==3.13.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pandas==2.1.3
numpy==1.24.3
geopandas==0.14.1
shapely==2.0.2
openai==1.3.0
python-dotenv==1.0.0
httpx==0.25.2